1대1 채팅 이미 존재시에도 반환타입 수정

+++ src/main/java/com/example/gazamung/mapper/ChatMapper.java
@@ -10,7 +10,7 @@ public interface ChatMapper {
     //    String findLatestMessageContentByChatRoomId(Long chatRoomId);
     Map<String, Object> findLatestMessageByChatRoomId(Long chatRoomId);
 
-    int countChatRoomsByMembersAndType(@Param("senderIdx") Long senderIdx, @Param("receiverIdx") Long receiverIdx, @Param("chatRoomType") int chatRoomType);
+    long countChatRoomsByMembersAndType(@Param("senderIdx") Long senderIdx, @Param("receiverIdx") Long receiverIdx, @Param("chatRoomType") int chatRoomType);
 
 
 }
